Add textHash2 column and PK (#2455)
* Added Primary Key

* Remove duplicates

* Use NullId constant during SearchParameterQueryGenerator

* Updated logging message

* Added Primary Key

* Adds pk for stringSearchParam table

* Update TextHash length

* Adds collation

* Addresses feedback

* Added TextHash2 column to TokenStringCompositeSearchParam

* Added comment

* Address feedback

* Added new types and row generators

* Updated Import BulkGenerators to use latest table types

+++ src/Microsoft.Health.Fhir.SqlServer/Features/Search/SqlSearchConstants.cs
@@ -9,6 +9,6 @@ namespace Microsoft.Health.Fhir.SqlServer.Features.Search
     {
         public const string SortSentinelValueForCt = "sentinelSortValue";
 
-        public const int NullId = 0;
+        public const short NullId = 0;
     }
 }
